Fix compilation with Xorg
Instead of using the old XFree86 directory assumptions, work with the
new Xorg way of putting the libraries at the standard /usr/lib, and
having the includes at /usr/include/X11. With this, there's no need for
an special parameter at configure to point to X11 libraries (as they'll
be find at /usr/lib).

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>