Fix xawtv DGA mode
v4l-conf needs a "base address". Normally it obtains it either from DGA
(under X), or from frame buffer (from fbtv). In my case a call to DGA
returns zero base address.

Without the patch it produces black window screen.

With patch applied, DGA mode works again, if V4L1 support is compiled
on the drivers (as it relies on Xorg "v4l" driver that it is V4L1 only).

The issue seems to be related to some problem with the X driver. The
relevant code at console/v4l-conf.c is:

    if (-1 == ioctl(fd, VC_GETMODE, (unsigned long)&mode)) {
        perror("ioctl VC_GETMODE");
        exit(1);
    }
    close(fd);
    d->width  = mode.width;
    d->height = mode.height;
    d->bpp    = mode.depth;
    d->bpl    = mode.pitch;
    d->base   = (void*)mode.fb_address;

It seems that mode.pitch is not returned with a correct value for some
X drivers. So, this patch is basically a workaround for a bug inside
the video adapter driver.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>