Add other XV attributes found on bttv + nv driver
This patch adds all attributes that can be managed with a bttv
board, using the nv driver, on a board that support overlay mode.

This change is not enough to make those controls visible, as
we would need to touch also at libng, but this is the first
step for doing it.

Overlay Xv mode is tested and is know to work, together with
the V4L Xorg driver.

Tested with a bttv Encore board and a FX5200 with the nv driver.
Nvidia chips up to NV30 family using the old nv driver are known
to work. Also, the old Radeon driver. Newer nvidia boards or
newer drivers like Nouveau don't support the old Xv overlay mode
anymore. A port to textured video is needed for overlay to
work on newer hardware/drivers.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>