Add optional support for libexplain
libexplain (http://libexplain.sourceforge.net/) provides an enhanced
description for the error messages, as it consolidates some knowledge
about each system call.

For example:
	ioctl: VIDIOC_QUERYMENU(id=134217731;index=2;name="Menu Item 1";reserved=0): ioctl(fildes = 4 "/dev/video0", request = _IORW('V', 37, int[11]), data = 0x7FFF361C2C20) failed, Invalid argument (22, EINVAL) because ioctl request or ioctl data is not valid
	ioctl: VIDIOC_REQBUFS(count=0;type=VIDEO_CAPTURE;memory=MMAP): ioctl(fildes = 4 "/dev/video0", request = _IORW('V', 8, int[5]), data = 0x01C8DA68) failed, Device or resource busy (16, EBUSY) because the character special device is in use by another process or by the system and this prevents the ioctl _IORW('V', 8, int[5]) system call from operating

If libexplain is available at the system, it will use for the V4L2 driver
(or libv4l). If not, it will use strerror() instead, as before.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>