alsa_stream: Improve common frequency find algorithm
The old algorithm started with the minimum frequency, which in case of
my radioSHARK resulted in a samplerate of only 9600 Hz, so start with
the highest rate. Also first try a couple of common-rates, to avoid ending
up with an unnecessary high rate, the first common rate tried is 44100 Hz,
to avoid unnecessary resampling as most outputs default to 44100.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>