common/alsa_stream: Improve common frequency algorithm yet again
Commit a9937e798ee8f565a66c6dda7cfd9e755081e407 changed the frequency
find algorithm to try a couple of common frequencies and then start at
the maximum rate, to avoid ending up with a too low rate, as seen on the
Griffin radioSHARK. But now with the analog input of the bt8xx I'm getting
the oposite, a too high rate. So modify the common frequency algorithm yet
again, to search upwards if the minimum support rate >= 44100 and downwards
otherwise.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>