Revert band support commits
It turns out I was a bit to quick with pushing support for the band selection
API, further review upstream has resulted in a significantly changed API,
so this reverts the band support commits as well as a few commits which
heavily depend on those changes. This will all return using the new API,
when the new band API has been merged upstream (so that we can be reasonably
sure it won't change again). Sorry about this...

This reverts commit 82cf6fc650f8fd9d2b28afc003cc4d5154381b09.
This reverts commit 0f326fa42ac0567908956af58698ecc627987de3.
This reverts commit da0ce71e376773989c9d9e97a27bfecc51ad3dca.
This reverts commit d728c3fafc848de527742783be8fac04db93e5f5.
This reverts commit f47c2c4f591c85e7388689066199428d6e3b6ec1.
This reverts commit 68a8c0610490a05b54e56f57fc23f5f412093501.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>