xawtv: Wait for our window to be configured
Before this patch it is possible for the main window to not be configured
yet when we call init_overlay.

When this happens video_gd_start exits without starting the stream because
video_gd_configure has not been called yet.

This leads to x11/xawtv: do_capture being called with an argument of
CAPTURE_OFF, which will then call x11_capture_pixmap -> ng_grabber_get_image
which does a read() on the /dev/video# device node, after which starting
an mmap stream or doing any further s_fmt calls will fail with EBUSY when
using drivers which don't allow mixing read and mmap (ie videobuf2 based
drivers).

Even with drivers which do allow the mixing and matching waiting is better
as it avoids unnecessary stream start/stop and format-renegotiation.

Note that this fix avoids xawtv sometimes not starting at all with drivers
which don't allow mixing read / mmap, but the snapshot functionality is still
broken with such drivers. I'll write a follow-up patch to fix this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>