drv0-v4l2: Fix snapshot function breaking streaming
Drivers using videobuf2 don't allow mixing and matching mmap and read calls,
so after the first read calls any s_fmt or reqbufs calls will return EBUSY,
thus taking a snapshot breaks the streaming (it does not resume after taking
the snapshot).

This patch fixes this by "detecting" these type of drivers by their reporting
of EBUSY on a s_fmt after a read call, and then re-opening the device (to
clear the EBUSY condition), and marking the device as not read() capable.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>