fbtv: allow not specifying the font full patchname
It is possible to use fbtv with a different font than lat1-16.
However, one has to place the full patchname. As we already
have a list of distro-specific patches and we do a search for
lat1-16, relax the logic to allow calling it with just the
font name, e. g.:
	fbtv -f ter-v32n

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>