console: allow specifying a font with its suffix
It is useful to allow specifying a font name with the suffix, as
this way, one could do something like:
	$ ls /usr/lib/kbd/consolefonts/

and pick a random file there, cutting-and-pasting the file
name. We could be smarter, but the font is loaded only once,
so performance is not an issue. So, just do it at the simplest
way.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>