fbtv: don't keep a black screen in case of fatal errors
If something bad happens and fbtv exits, it should restore
the VT window, as otherwise the error messages won't be
displayed.

So, be sure to call fb_cleanup() before calling exit().

Let's also start ncurses later, as this is used only to get
keys, and this happens only within the image display loop.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>