libng: get rid of VIDIOC_G_PARM ioctl
This ioctl is used only to enumerate stuff, and at the wrong
place. The common usage for this syscall is to get the
frames per second rate of a video stream, but that only works
after starting the stream.

Also, despite getting the information, xawtv never uses it,
not even to display the obtained values.

So, just get rid of it.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>